Report error in Ibex run target
The current output for an error in the execution of the Ibex target is
not reported. All tests are executed and at the verification stage a
difference in the output signatures will result in a failure of the test
case.
As no error of `RUN_TARGET` is produced, it is difficult to investigate
the actual problem.

This change will allow `make` to report an error but continue the
execution of the following tests.
As the output of a test case now shows that an error occurred, it is
easier to find the cause for a failed test if an error happened in the
execution of the target.

Closes riscv/riscv-compliance#70