Adding RVMODEL_CHECK_TEST.  Putting back RVMODEL_IO_ASSERT_GPR_EQ
New macro will check the test output based on
expected value from the test and correctval.

This will basically compare what has been written to
the signature block vs what is expected there.  It is not
ment as a replacement for the verify step after the tests are
complete to compare the signature block to the reference_output.

It is ment to be used as a debug or sanity check for:
1. That the tests are still running and have not locked up
2. During development to see what if any tests fail/pass

The RVMODEL_CHECk_TEST() macro is not ment to stop execution on
a fail conditon.

The RVMODEL_IO_ASSERT_GPR_EQ() macro can be used to stop execution
on a fail condition.

Both of these are suggested uses for these macros as the implementation
of the macros is dependent on the target development. They are defined
in the compliance_model.h file in the respective riscv-target/<target> dir.