Update zext.h-01.S
Remove RV64IB from the ISA string.

The RV64IB ISA causes a compilation string of -march=rv64ib and thus a gcc error "cannot find default versions of the ISA extension 'b'".

This is the only test in riscv-arch-test that specifies RV64IB.  I believe it is a typo and should be removed.  When I remove it, riscof produces a Makefile that compiles zext.h-01 successfully.

Signed-off-by: David Harris <74973295+davidharrishmc@users.noreply.github.com>