Fix undefined label reference in TEST_JAL_OP
When assembling with LLVM, this macro produces the following:
`error: directional label undefined`. This happens because it compares
the symbol value rather than performing a textual comparison. At the
point the 3f label is referenced it is actually a backwards reference
so we would have to use 3b instead. Fix this by using a `.ifc` string
comparison directive instead.