Add menvcfg support (Zicbom_Zicboz_Zicsr_Ssdtso_Sstc)
The {m,s,h}envcfg CSRs have been defined a long time ago.
They contain bits that are defined by a the following ISA extensions:
  Zicbom_Zicboz_Zicsr_Ssdtso_Sstc

This patch introduces tests which read and write all of the defined
bitsfields of these registers from M, S, and U mode.

In case of (expected) valid accesses we toggle the bits twice to test
if we can set and clear the bits.
In case of (expected) illegal accesses we capture the exception state
in the signature using the default ACT trap handler code.

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>